requirements.txt
langchain_OpenAI
langchain_core
python-dotenv
streamlit
langchain_community